requirements.txt
aniso8601==6.0.0
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.0
Werkzeug==0.14.1
